More optional ethcore features: price-info, stratum
