+++ ethcore/src/miner/mod.rs
@@ -23,6 +23,7 @@ mod miner;
 mod service_transaction_checker;
 
 pub mod pool_client;
+#[cfg(feature = "stratum")]
 pub mod stratum;
 
 pub use self::miner::{Miner, MinerOptions, Penalization, PendingSet, AuthoringParams};
