+++ miner/Cargo.toml
@@ -26,7 +26,7 @@ keccak-hash = { path = "../util/hash" }
 linked-hash-map = "0.5"
 log = "0.3"
 parking_lot = "0.6"
-price-info = { path = "../price-info" }
+price-info = { path = "../price-info", optional = true }
 rlp = { path = "../util/rlp" }
 trace-time = { path = "../util/trace-time" }
 transaction-pool = { path = "../transaction-pool" }
